extract messageFormatFactory to lib folder
